logic improvements
